+++ packages/tools/geo/index.js
@@ -1,2 +1,3 @@
-export * from './src/modules/geojson';
-export * from './src/modules/topojson';
+export * from './src/modules/geojson.js';
+export {default as pruneTopology} from './src/modules/pruneTopology.js';
+export * from './src/modules/topojson.js';